requirements.txt
# --------- pytorch --------- #
#torch>=2.0.0
#torchvision>=0.15.0
#torchaudio>=0.13.01

# -------- lightning -------- #
lightning>=2.0.0
pytorch-lightning==1.9.3
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
omegaconf

# -------- dataset -------- #
librosa==0.9.1
torchcrepe>=0.0.16
h5py>=3.7.0

# ------ state space ------- #
numpy
scipy
pandas
scikit-learn
matplotlib
tqdm
rich
lit # Getting installation errors with torch 2.0 if this isn't installed
einops
opt-einsum
cmake # For pykeops support
# pykeops # Seems to cause various issues; leaving uninstalled by default
# transformers # For some schedulers

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
ipykernel       # for notebooks
# sh            # for running bash commands in some tests (linux/macos only)

